requirements.txt
Flask==2.1.0
flask-restful
flask-Cors
Jinja2
python-dotenv
gunicorn
uvicorn
line-bot-sdk
requests